Solution to day 7
The code needs cleanup, improve the FS implementation and add tests.

The FS tree is created *almost* correctly to get the solutions to the problems, but some of the directories are repeated with size 0. That means it works for solving the puzzles (get the sum of the node sizes and find the minimum directory that frees up the required space), but they're not the best implementation.

Remove unneded file